feat: add profile stack
